[feat] #11-FAQ 최종(QA+Feedback)
