--- src/start/FAQ_QA.css
@@ -1,90 +1,0 @@
-.faq-page {
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    width: 393px;
-    height: 852px;
-    margin: 0 auto;
-    background: #ffffff;
-    font-family: "Inter", sans-serif;
-    overflow-y: scroll;
-    height: 852px;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-
-  .faq-tabs {
-    display: flex;
-    justify-content: left;
-    align-items: center;
-    margin-top: 10px;
-  }
-  
-  .faq-tab {
-    font-size: 12px;
-    font-weight: 700;
-    color: #422919;
-    border: none;
-    border-radius: 0;
-    background: #ffffff;
-    cursor: pointer;
-  }
-  
-  .faq-tab.active {
-    border-bottom: 2.5px solid #422919;
-  }
-  
-  /* FAQ 리스트 */
-  .faq-list {
-    margin-top: 40px;
-    padding: 0 25px;
-  }
-  
-  .faq-item {
-    border-bottom: 1px solid #e3e3e3;
-    padding: 15px 0;
-    margin-bottom: 10px;
-  }
-  
-  .faq-question {
-    display: flex;
-    justify-content: space-between;
-    align-items: left;
-    cursor: pointer;
-    text-align: left;
-  }
-  
-  .faq-category {
-    margin-left: 5px;
-    margin-right: 20px;
-    font-size: 12px;
-    font-weight: 700;
-    color: #422919;
-    width: 70px;
-    flex-shrink: 0;
-  }
-  
-  .faq-q {
-    flex-grow: 1;
-    font-size: 11px;
-    font-weight: 500;
-    color: #2b2b2b;
-    padding-left: 5px;
-  }
-  
-  .faq-toggle-icon {
-    width: 20px;
-    height: 20px;
-    margin-bottom: 5px;
-    margin-right: 5px;
-  }
-  
-  .faq-answer {
-    font-size: 11px;
-    font-weight: 400;
-    color: #2b2b2b;
-    margin-top: 10px;
-    padding-left: 100px;
-    text-align: left;
-  }
